Fixes Middleware unit tests where stake to power is used.
Moves setting of vault collateral to new hook after register vault.
There is a check in place for the vault being valid which happens on
register call, if you try to get the colleteral on the before hook,
and the vault is not valid, you may get unexpected revert.

Setting the collateral in the after register hook prevents this issue
and avoids double check of valid vault.